Added Worker Map checks

+++ core/required/daemon.js
@@ -105,21 +105,29 @@ module.exports = (function() {
 
       err = err || new Error('Application Stopped');
 
-      // We only have one worker for now so this is rather easy
-      let mapIterator = this._workers.value();
-      let worker = mapIter.next().value;
-
-      worker.send({ __destroy__: true });
-      worker.once('message', (msg) => {
-        if ((msg) && (typeof msg === 'object') && (msg.__destroy__ && msg.ready)) {
-          
-          worker.kill();
-          worker.once('exit', (code, signal) => {
-            onStop();
-          });
-          
-        }
-      });
+      // Make sure we have a Worker
+      if (this._workers.size) {
+        // We only have one worker for now so this is rather easy
+        let mapIterator = this._workers.value();
+        let worker = mapIter.next().value;
+        
+        worker.send({ __destroy__: true });
+        worker.once('message', (msg) => {
+          if ((msg) && (typeof msg === 'object') && (msg.__destroy__ && msg.ready)) {
+            
+            worker.kill();
+            worker.once('exit', (code, signal) => {
+              onStop(() => {});
+            });
+            
+          }
+        });
+        
+      }else{
+        
+        onStop(() => {});
+        
+      }
       
     }
